RunReportRequest Object

RunReportRequest Object

RunReportRequest Object

RunReportRequest(
  entity = NULL,
  metrics = NULL,
  currencyCode = NULL,
  orderBys = NULL,
  metricFilter = NULL,
  cohortSpec = NULL,
  metricAggregations = NULL,
  dimensions = NULL,
  dateRanges = NULL,
  keepEmptyRows = NULL,
  dimensionFilter = NULL,
  limit = NULL,
  offset = NULL,
  returnPropertyQuota = NULL
)

RunReportRequest(
  entity = NULL,
  metrics = NULL,
  currencyCode = NULL,
  orderBys = NULL,
  metricFilter = NULL,
  cohortSpec = NULL,
  metricAggregations = NULL,
  dimensions = NULL,
  dateRanges = NULL,
  keepEmptyRows = NULL,
  dimensionFilter = NULL,
  limit = NULL,
  offset = NULL,
  returnPropertyQuota = NULL
)

Arguments

entity
A property whose events are tracked

metrics
The metrics requested and displayed

currencyCode
A currency code in ISO4217 format, such as 'AED', 'USD', 'JPY'

orderBys
Specifies how rows are ordered in the response - see ga_data_order

metricFilter
The filter clause of metrics - see ga_data_filter

cohortSpec
Cohort group associated with this request

metricAggregations
Aggregation of metrics. Aggregated metric values will be shown in rows where the dimensionValues are set to "RESERVED_(MetricAggregation)".

dimensions
The dimensions requested and displayed

dateRanges
Date ranges of data to read, can send up to 4 date ranges (a vector of 8 dates spanning 1 year)

keepEmptyRows
If false, rows with metrics being 0 will not be returned

dimensionFilter
The filter clause of dimensions - see ga_data_filter

limit
The number of rows to return - use -1 to return all rows

offset
The row count of the start row

returnPropertyQuota
Toggles whether to return the current state of this Analytics Property's

Value

RunReportRequest object

RunReportRequest object

Details

Autogenerated via gar_create_api_objects The request to generate a report.

Autogenerated via gar_create_api_objects The request to generate a report.