Aggregate a Google Analytics dataframe over inputted columns

A helper function to aggregate over dimensions

aggregateGAData(ga_data, agg_names = NULL,
  mean_regex = "^avg|^percent|Rate$|^CPC$|^CTR$|^CPM$|^RPC$|^ROI$|^ROAS$|Per")

Arguments

ga_data
A dataframe of data to aggregate

agg_names
The columns to aggregate over

mean_regex
The regex for column names to do mean() rather than sum()

Details

Will auto select metrics if they are numeric class columns. Will auto perform mean aggregation it metric names match mean_regex argument If agg_names is NULL will aggregate over all

Examples

# NOT RUN {
# use `aggregateGAData` so you can on the fly create summary data
ga_data <- google_analytics(81416156,
                            date_range = c("10daysAgo", "yesterday"),
                            metrics = "sessions", dimensions = c("hour","date"))

# if we want totals per hour over the dates:
aggregateGAData(ga_data[,c("hour","sessions")], agg_names = "hour")

# it knows not to sum metrics that are rates:
aggregateGAData(ga_data[,c("hour","bounceRate")], agg_names = "hour")

# }