Create named list of allowed GA metrics/dimensions

Create named list of allowed GA metrics/dimensions

allowed_metric_dim(type = c("METRIC", "DIMENSION"), subType = c("all",
  "segment", "cohort"), callAPI = FALSE)

Arguments

type

Type of parameter to create

subType

to restrict to only those in this type

callAPI

This will update the meta table (Requires online authorization) This is useful to expand goalXCompletions to all the possiblilties, as well as restricting to those that variables that work with your API call. Use internal meta table, but you have option to update to the latest version.

Value

A named list of parameters for use in API calls