Make a dimension or metric filter clause object

Make a dimension or metric filter clause object

filter_clause_ga4(filters, operator = c("OR", "AND"))

Arguments

filters
a list of dim_filter or met_filter. Only one type allowed.

operator
combination of filter.

Value

An object of class dim_fil_ga4 or met_fil_ga4 for use in make_ga_4_req

Details

If you have dimension and metric filters, make the clauses in two separate calls, then pass the objects to make_ga_4_req

See also

Examples

if (FALSE) {
library(googleAnalyticsR)

## authenticate,
## or use the RStudio Addin "Google API Auth" with analytics scopes set
ga_auth()

## get your accounts
account_list <- google_analytics_account_list()

## pick a profile with data to query

ga_id <- account_list[23,'viewId']

## create filters on metrics
mf <- met_filter("bounces", "GREATER_THAN", 0)
mf2 <- met_filter("sessions", "GREATER", 2)

## create filters on dimensions
df <- dim_filter("source","BEGINS_WITH","1",not = TRUE)
df2 <- dim_filter("source","BEGINS_WITH","a",not = TRUE)

## construct filter objects
fc2 <- filter_clause_ga4(list(df, df2), operator = "AND")
fc <- filter_clause_ga4(list(mf, mf2), operator = "AND")

## make v4 request
ga_data1 <- google_analytics(ga_id,
                             date_range = c("2015-07-30","2015-10-01"),
                             dimensions=c('source','medium'),
                             metrics = c('sessions','bounces'),
                             met_filters = fc,
                             dim_filters = fc2,
                             filtersExpression = "ga:source!=(direct)")

}