Delete a filter from account or remove from view.

Delete a filter from account or remove from view.

ga_filter_delete(
  accountId,
  webPropertyId = NULL,
  viewId = NULL,
  filterId,
  removeFromView = FALSE
)

Arguments

accountId

Account Id of the account that contains the filter

webPropertyId

Property Id of the property that contains the filter

viewId

View Id of the view that contains the filter

filterId

Filter Id of the filter to be deleted

removeFromView

Default if FALSE. If TRUE, deletes the filter from the view

Value

TRUE if successful

See also