Create or update user access to Google Analytics

If you supply more than one email, then batch processing will be applied. Batching has special rules that give you 30 operations for the cost of one API call against your quota. When batching you will only get a TRUE result on successful batch, but individual entries may have failed. Check via ga_users_list afterwards and try to add individual linkIds to get more descriptive error messages.

ga_users_add(
  email,
  permissions,
  accountId,
  webPropertyId = NULL,
  viewId = NULL
)

Arguments

email
The email(s) of the user(s) to add. Has to have a Google account.

permissions
Which permissions to add as a vector - "MANAGE_USERS","EDIT","COLLABORATE","READ_AND_ANALYZE"

accountId
Account Id

webPropertyId
Web Property Id - set to NULL to operate on account level only

viewId
viewId - set to NULL to operate on webProperty level only

Value

TRUE if successful

See also

Examples

if (FALSE) {
library(googleAnalyticsR)
ga_auth()

ga_users_add(c("the_email@company.com", "another_email@company.com"), 
             permissions = "EDIT", accountId = 47480439)

}