Library for getting Google Analytics data into R

Follow the online documentation here: http://code.markedmondson.me/googleAnalyticsR/

Details

First Google Analytics Reporting v4 API library for R

v4 features include: dynamic calculated metrics, pivots, histograms, date comparisons, batching

Multi-user login in Shiny App to enable Google Analytics powered Shiny Apps

Integration with BigQuery Google Analytics Premium/360 exports

Single authentication flow with other googleAuthR apps

Automatic batching

Multi-channel funnnels

Support for batch. For big data calls this could be 10x quicker than normal GA fetching

Meta data included in attributes of returned dataframe

Automatic caching of API requests

Management API features such as unsampled report downloads and view filter setup

You may wish to set the below environment arguments for easier authentication

GA_CLIENT_ID GA_CLIENT_SECRET GA_WEB_CLIENT_ID GA_WEB_CLIENT_SECRET GA_AUTH_FILE