Make a Measurement Protocol v2 request

Create a server side call to Google Analytics 4 via its Measurement Protocol

ga_mp_send(
  events,
  client_id,
  measurement_id,
  user_id = NULL,
  api_secret = Sys.getenv("GA_MP_SECRET"),
  debug = FALSE,
  timestamp_micros = NULL,
  user_properties = NULL,
  non_personalized_ads = TRUE
)

Arguments

events
The events to send

client_id
The client_id to associate with the event

measurement_id
The measurement ID associated with a stream

user_id
Optional. Unique id for the user

api_secret
The secret generated in the GA4 UI - by default will look for environment arg GA_MP_SECRET

debug
Send hits to the debug endpoint to validate hits

timestamp_micros
Optional. A Unix timestamp (in microseconds) for the time to associate with the event.

user_properties
Optional. The user properties for the measurement sent in as a named list.

non_personalized_ads
Optional. Set to true to indicate these events should not be used for personalized ads.

Value

TRUE if successful, if debug=TRUE then validation messages if not a valid hit.

Details

Create an API secret via Admin > Data Streams > choose your stream > Measurement Protocol > Create

To see event parameters, create custom fields in your GA4 account first, to see them in your reports 24hrs after you send them in with this function via Custom definitions > Create custom dimensions - dimension name will be how it looks like in the reports, event parameter will be the parameter you have sent in with the event.

user_id can be used for cross-platform analysis

timestamp_micros should only be set to record events that happened in the past. This value can be overridden via user_property or event timestamps. Events can be backdated up to 48 hours. Note microseconds, not milliseconds.

user_properties - describe segments of your user base, such as language preference or geographic location. See User properties

Ensure you also have user permission as specified in the feature policy

Invalid events are silently rejected with a 204 response, so use debug=TRUE to validate your events first.

See also

Examples

# preferably set this in .Renviron
Sys.setenv(GA_MP_SECRET="MY_SECRET")

# your GA4 settings
my_measurement_id <- "G-43MDXK6CLZ"
a_client_id <- 1234567

event <- ga_mp_event("an_event")
ga_mp_send(event, a_client_id, my_measurement_id, debug = TRUE)

#> ℹ 2021-03-23 22:52:40 > MP Request: https://www.google-analytics.com/debug/mp/collect?measurement_id=G-43MDXK6CLZ&api_secret=MY_SECRET 
#>  {
#>   "client_id": "1234567",
#>   "non_personalized_ads": true,
#>   "events": {
#>     "name": "an_event"
#>   }
#> }

#> ℹ 2021-03-23 22:52:40 > Response:  200

#> ℹ 2021-03-23 22:52:40 > No validation messages found

#> [1] TRUE

another <- ga_mp_event("another_event")
ga_mp_send(list(event, another), 
           a_client_id, my_measurement_id, 
           debug = TRUE)

#> ℹ 2021-03-23 22:52:40 > MP Request: https://www.google-analytics.com/debug/mp/collect?measurement_id=G-43MDXK6CLZ&api_secret=MY_SECRET 
#>  {
#>   "client_id": "1234567",
#>   "non_personalized_ads": true,
#>   "events": [
#>     {
#>       "name": "an_event"
#>     },
#>     {
#>       "name": "another_event"
#>     }
#>   ]
#> }

#> ℹ 2021-03-23 22:52:40 > Response:  200

#> ℹ 2021-03-23 22:52:40 > No validation messages found

#> [1] TRUE

if (FALSE) {
# you can see sent events in the real-time reports
my_property_id <- 206670707
ga_data(my_property_id, 
        dimensions = "eventName", 
        metrics = "eventCount", 
        dim_filters = ga_data_filter(
           eventName == c("an_event","another_event")),
        realtime = TRUE)

}